+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Features
 
 * Add #PAC-85
+* Add functionality to also parse configuration files that'll be available in configured additional vendor directory
 * Update to versions of techdivision/import 16.*, techdivision/import-configuration-jms 16.* and techdivision/import-app-simple 16.*
 
 # Version 9.0.4